// src/app/core/models/empresa.model.ts
export interface Empresa {
  MA04_Unidad: string;
  MA04_Nombre: string;
  MA04_RazonSocial?: string;
  MA04_Direccion?: string;
  MA04_Ruc?: string;
  MA04_Ciudad?: string;
  MA04_Pais?: string;
  MA04_Telefono1?: string;
  MA04_Email?: string;
  MA04_Web?: string;
  MA04_Principal?: number;
  MA04_Activo?: number;
}